add color map to use different colors for different objects
